Move pytest configure to conftest
