+++ src/shipchain_common/test_utils/helpers.py
@@ -134,61 +134,6 @@ class GeoCoderResponse:
         self.xy = point
 
 
-def pytest_configure():
-    MIDDLEWARE = (
-        'django.middleware.common.CommonMiddleware',
-        'django.contrib.sessions.middleware.SessionMiddleware',
-        'django.contrib.auth.middleware.AuthenticationMiddleware',
-        'django.contrib.messages.middleware.MessageMiddleware',
-    )
-
-    settings.configure(
-        DEBUG_PROPAGATE_EXCEPTIONS=True,
-        DATABASES={
-            'default': {
-                'ENGINE': 'django.db.backends.sqlite3',
-                'NAME': ':memory:'
-            }
-        },
-        SITE_ID=1,
-        SECRET_KEY='not very secret in tests',
-        USE_I18N=True,
-        USE_L10N=True,
-        STATIC_URL='/static/',
-        ROOT_URLCONF='tests.urls',
-        TEMPLATES=[
-            {
-                'BACKEND': 'django.template.backends.django.DjangoTemplates',
-                'APP_DIRS': True,
-            },
-        ],
-        MIDDLEWARE=MIDDLEWARE,
-        MIDDLEWARE_CLASSES=MIDDLEWARE,
-        INSTALLED_APPS=(
-            'django.contrib.auth',
-            'django.contrib.contenttypes',
-            'django.contrib.sessions',
-            'django.contrib.sites',
-            'django.contrib.staticfiles',
-
-            'rest_framework',
-            'rest_framework_simplejwt',
-            'rest_framework_simplejwt.token_blacklist',
-
-            'tests',
-        ),
-        PASSWORD_HASHERS=(
-            'django.contrib.auth.hashers.MD5PasswordHasher',
-        ),
-    )
-
-    try:
-        import django
-        django.setup()
-    except AttributeError:
-        pass
-
-
 # pylint: disable=too-many-arguments, redefined-builtin
 class JsonAPIClient(APIClient):
     def post(self, path, data=None, format='json', content_type=None, follow=False, **extra):
